--- README.md
@@ -1,2 +1,0 @@
-# AndroidTemplate
-Personal template for all Android projects
