+++ app/build.gradle.kts
@@ -1,14 +1,30 @@
+/*
+ * Copyright (C) 2022, Alyssa Nash <nash.alyssab@gmail.com>
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
 plugins {
     id("com.android.application")
     id("org.jetbrains.kotlin.android")
 }
 
 android {
-    namespace = "com.syriusdevelopment.androidtemplate"
+    namespace = "com.syriusdevelopment.atlas"
     compileSdk = 33
 
     defaultConfig {
-        applicationId = "com.syriusdevelopment.androidtemplate"
+        applicationId = "com.syriusdevelopment.atlas"
         minSdk = 23
         targetSdk = 33
         versionCode = 1
