+++ app/src/main/java/com/syriusdevelopment/atlas/MainActivity.kt
@@ -1,4 +1,20 @@
-package com.syriusdevelopment.androidtemplate
+/*
+ * Copyright (C) 2022, Alyssa Nash <nash.alyssab@gmail.com>
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+package com.syriusdevelopment.atlas
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -10,13 +26,13 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import com.syriusdevelopment.androidtemplate.ui.theme.AndroidTemplateTheme
+import com.syriusdevelopment.atlas.ui.theme.AtlasTheme
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContent {
-            AndroidTemplateTheme {
+            AtlasTheme {
                 // A surface container using the 'background' color from the theme
                 Surface(
                     modifier = Modifier.fillMaxSize(),
@@ -37,7 +53,7 @@ fun Greeting(name: String, modifier: Modifier = Modifier) {
 @Preview(showBackground = true)
 @Composable
 fun DefaultPreview() {
-    AndroidTemplateTheme {
+    AtlasTheme {
         Greeting("Android")
     }
 }
