+++ app/src/main/java/com/syriusdevelopment/atlas/ui/theme/Theme.kt
@@ -1,4 +1,20 @@
-package com.syriusdevelopment.androidtemplate.ui.theme
+/*
+ * Copyright (C) 2022, Alyssa Nash <nash.alyssab@gmail.com>
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+package com.syriusdevelopment.atlas.ui.theme
 
 import android.app.Activity
 import android.os.Build
@@ -38,7 +54,7 @@ private val LightColorScheme = lightColorScheme(
 )
 
 @Composable
-fun AndroidTemplateTheme(
+fun AtlasTheme(
     darkTheme: Boolean = isSystemInDarkTheme(),
     // Dynamic color is available on Android 12+
     dynamicColor: Boolean = true,
