+++ lint/copyrights/copyright.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2022, Alyssa Nash <nash.alyssab@gmail.com>
+ * Copyright (C) $YEAR, Alyssa Nash <nash.alyssab@gmail.com>
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -14,19 +14,3 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-pluginManagement {
-    repositories {
-        gradlePluginPortal()
-        google()
-        mavenCentral()
-    }
-}
-dependencyResolutionManagement {
-    repositoriesMode.set(RepositoriesMode.FAIL_ON_PROJECT_REPOS)
-    repositories {
-        google()
-        mavenCentral()
-    }
-}
-rootProject.name = "Atlas"
-include(":app")
