+++ lint/copyrights/copyright.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <!--
-    Copyright (C) 2022, Alyssa Nash <nash.alyssab@gmail.com>
+    Copyright (C) $YEAR, Alyssa Nash <nash.alyssab@gmail.com>
 
     This program is free software: you can redistribute it and/or modify
     it under the terms of the GNU General Public License as published by
@@ -15,7 +15,3 @@
     You should have received a copy of the GNU General Public License
     along with this program.  If not, see <http://www.gnu.org/licenses/>.
 -->
-<resources>
-
-    <style name="Theme.Atlas" parent="android:Theme.Material.Light.NoActionBar" />
-</resources>
